Moving error constants to existing Constants class
Moving error constants to existing Constants class
