--- components/org.wso2.carbon.identity.oauth.dcr/src/main/java/org/wso2/carbon/identity/oauth/dcr/DCRConfigErrorMessage.java
@@ -1,98 +1,0 @@
-/*
- * Copyright (c) 2024, WSO2 LLC. (http://www.wso2.com).
- *
- * WSO2 LLC. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.wso2.carbon.identity.oauth.dcr;
-
-/**
- * Error Message class for DCR Configuration errors.
- */
-public enum DCRConfigErrorMessage {
-
-    /**
-     * Invalid tenant domain.
-     */
-    ERROR_CODE_INVALID_TENANT_DOMAIN("60005",
-            "Invalid input.",
-            "%s is not a valid tenant domain."),
-
-    /**
-     * Unable to retrieve DCR configuration.
-     */
-    ERROR_CODE_DCR_CONFIGURATION_RETRIEVE("65020",
-            "Unable to retrieve DCR configuration.",
-            "Server encountered an error while retrieving the " +
-                    "DCR configuration of %s.");
-
-    /**
-     * The error code.
-     */
-    private final String code;
-
-    /**
-     * The error message.
-     */
-    private final String message;
-
-    /**
-     * The error description.
-     */
-    private final String description;
-
-
-    DCRConfigErrorMessage(String code, String message, String description) {
-        this.code = code;
-        this.message = message;
-        this.description = description;
-    }
-
-    /**
-     * Get the {@code code}.
-     *
-     * @return Returns the {@code code} to be set.
-     */
-    public String getCode() {
-
-        return code;
-    }
-
-    /**
-     * Get the {@code message}.
-     *
-     * @return Returns the {@code message} to be set.
-     */
-    public String getMessage() {
-
-        return message;
-    }
-
-    /**
-     * Get the {@code description}.
-     *
-     * @return Returns the {@code description} to be set.
-     */
-    public String getDescription() {
-
-        return description;
-    }
-
-    @Override
-    public String toString() {
-
-        return code + ":" + message;
-    }
-}
