+++ components/org.wso2.carbon.identity.oauth.dcr/src/main/java/org/wso2/carbon/identity/oauth/dcr/DCRConfigurationMgtServiceImpl.java
@@ -69,7 +69,8 @@ public class DCRConfigurationMgtServiceImpl implements DCRConfigurationMgtServic
         try {
             IdentityTenantUtil.getTenantId(tenantDomain);
         } catch (IdentityRuntimeException e) {
-            throw handleClientException(DCRConfigErrorMessage.ERROR_CODE_INVALID_TENANT_DOMAIN, e, tenantDomain);
+            throw handleClientException(DCRMConstants.DCRConfigErrorMessage.ERROR_CODE_INVALID_TENANT_DOMAIN, e,
+                    tenantDomain);
         }
     }
